+++ package.json
@@ -12,7 +12,6 @@
     "axios": "^1.6.5",
     "js-cookie": "^3.0.5",
     "react": "^18.2.0",
-    "react-cookie": "^7.0.1",
     "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
     "react-router-dom": "^6.21.3",
