--- src/assets/styles/components/Post.css
@@ -1,17 +1,0 @@
-::-webkit-scrollbar {
-    width: 7px;
-  }
-  
-  ::-webkit-scrollbar-track {
-    background: rgb(238 242 255);
-    border-radius: 20px;
-  }
-
-  ::-webkit-scrollbar-thumb {
-    background: rgb(99 102 241);
-    border-radius: 20px;
-  }
-  
-  ::-webkit-scrollbar-thumb:hover {
-    background: rgb(67 56 202);
-  }
