+++ src/components/CustomButton.jsx
@@ -1,0 +1,11 @@
+import Button from "@mui/material/Button";
+
+const CustomButton = ({ children, type, onClick }) => {
+  return (
+    <Button variant="contained" type={type} onClick={onClick}>
+      {children}
+    </Button>
+  );
+};
+
+export default CustomButton;
