+++ src/pages/login/index.jsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import toast, { Toaster } from 'react-hot-toast';
 
-import Input from "../../common/Input";
-import ButtonComponent from "../../common/ButtonComp";
-import { logIn } from "../../../services/authService";
+import Input from "../../components/Input";
+import CustomButton from "../../components/CustomButton";
+import { logIn } from "../../services/authService";
 
 const Login = () => {
   const navigate = useNavigate();
@@ -14,12 +14,9 @@ const Login = () => {
     password: "",
   });
 
-  const isNull = () => {
+  const isEmpty = () => {
     const { email, password } = data;
-    if (email === "" || password === "") {
-      return true;
-    }
-    return false;
+    return (email === "" || password === "")
   };
 
   const handleChange = (e) => {
@@ -34,31 +31,31 @@ const Login = () => {
 
   const handleSubmit = async (e) => {
     e.preventDefault();
-    if (!isNull()) {
-      const res = await logIn(data);
-      if (!res.error) {
-        toast.success(res.message)
-        navigate("/dashboard");
-      } else {
-        toast.error(res.error)
-      }
-    } else {
-      toast.error("Please fill all the values!")
+    if (isEmpty()) {
+      toast.error("Please fill all the values!");
+      return;
     }
+    const res = await logIn(data);
+    if (res.data.error || res.status !== 200) {
+      toast.error(res.data.error)
+      return;
+    }
+    toast.success(res.data.message)
+    navigate("/dashboard");
   };
 
   const inputs = [
     {
       name: "email",
       type: "email",
-      label: "email",
+      label: "Email",
       placeholder: "example@gmail.com",
       value: data.email,
     },
     {
       name: "password",
       type: "password",
-      label: "password",
+      label: "Password",
       placeholder: "Password",
       value: data.password,
     },
@@ -67,13 +64,9 @@ const Login = () => {
   return (
     <>
       <div className="flex justify-center items-center h-[100vh] bg-blue-100">
-        <form
-          method="POST"
-          action="/user/signin"
-          className="p-5 rounded-lg bg-blue-50"
-        >
+        <form method="POST" className="p-5 rounded-lg bg-blue-50">
           <p className="justify-center flex text-lg text-blue-600 font-bold">
-            LogIn
+            Log in
           </p>
           <div className="flex flex-col justify-around min-h-[35vh] w-[20rem]">
             {inputs.map((input, index) => {
@@ -92,13 +85,13 @@ const Login = () => {
               );
             })}
             <div className="flex justify-start">
-              <ButtonComponent type="submit" onClick={handleSubmit}>
-                Login
-              </ButtonComponent>
+              <CustomButton type="submit" onClick={handleSubmit}>
+                Log in
+              </CustomButton>
               <span className=" ml-2 flex my-auto text-sm">
-                Don't have an account?
-                <Link to="/signup" className="text-blue-500">
-                  SignUp
+                Don't have an account?&nbsp;
+                <Link to="/signup" className="text-blue-500 font-bold">
+                  Sign up
                 </Link>
               </span>
             </div>
